+++ packages/components/src/overlayHeader/OverlayHeader.js
@@ -5,6 +5,7 @@ import Header from '../header';
 import CloseButton from '../common/closeButton';
 
 import './OverlayHeader.css';
+import Logo from '../logo';
 
 const OverlayHeader = ({ avatar, onClose, logo }) => {
   const closeButton = onClose && <CloseButton onClick={onClose} />;
@@ -25,13 +26,14 @@ const OverlayHeader = ({ avatar, onClose, logo }) => {
 
 OverlayHeader.defaultProps = {
   avatar: null,
+  logo: <Logo />,
   onClose: null,
 };
 
 OverlayHeader.propTypes = {
   /** An Avatar */
   avatar: PropTypes.element,
-  logo: PropTypes.node.isRequired,
+  logo: PropTypes.node,
   /** Function called when the close is clicked */
   onClose: PropTypes.func,
 };