+++ packages/components/src/overlayHeader/OverlayHeader.story.js
@@ -4,8 +4,8 @@ import { action } from '@storybook/addon-actions';
 import { Profile as ProfileIcon, Briefcase as BriefcaseIcon } from '@transferwise/icons';
 import OverlayHeader from './OverlayHeader';
 import Avatar from '../avatar';
-import demoLogo from '../../public/assets/logo_full.svg';
 import AvatarWrapper from '../common/avatarWrapper';
+import Logo from '../logo';
 
 export default {
   component: OverlayHeader,
@@ -22,7 +22,7 @@ export const basic = () => {
   const showAvatar = select('avatar', Object.keys(avatarProfiles), 'Profile');
   return (
     <OverlayHeader
-      logo={<img alt="logo" src={demoLogo} width="138" />}
+      logo={<Logo />}
       onClose={action('Close clicked')}
       avatar={
         <Avatar type={Avatar.Type.ICON} size={Avatar.Size.MEDIUM}>
@@ -38,7 +38,7 @@ export const withAvatarWrapper = () => {
   const profileType = select('profileType', Object.keys(AvatarWrapper.ProfileType));
   return (
     <OverlayHeader
-      logo={<img alt="logo" src={demoLogo} width="138" />}
+      logo={<Logo />}
       onClose={action('Close clicked')}
       avatar={<AvatarWrapper url={avatarURL} profileType={profileType} />}
     />