+++ packages/docs/liveEditorCode/logo.code.js
@@ -1,0 +1,1 @@
+<Logo type={LogoType.WISE} />;