+++ packages/dynamic-flows/src/flow/layoutService/LayoutService.spec.js
@@ -1,4 +1,4 @@
-import { convertStepToLayout, inlineFormSchemas } from '.';
+import { convertStepToLayout, inlineReferences } from '.';
 
 describe('Given a utility service for handling dynamic layouts', () => {
   describe('when we receive a decision step', () => {
@@ -115,7 +115,7 @@ describe('Given a utility service for handling dynamic layouts', () => {
             {
               type: 'button',
               context: exitAction.type,
-              action: { ...exitAction, type: undefined },
+              action: { ...exitAction, label: exitAction.title, title: undefined, type: undefined },
             },
           ],
         },
@@ -214,12 +214,12 @@ describe('Given a utility service for handling dynamic layouts', () => {
             {
               type: 'button',
               context: submitAction.type,
-              action: { ...submitAction, type: undefined },
+              action: { ...submitAction, label: submitAction.title, title: undefined, type: undefined },
             },
             {
               type: 'button',
               context: cancelAction.type,
-              action: { ...cancelAction, type: undefined },
+              action: { ...cancelAction, label: cancelAction.title, title: undefined, type: undefined },
             },
           ],
         },
@@ -229,7 +229,7 @@ describe('Given a utility service for handling dynamic layouts', () => {
     });
   });
 
-  describe('when asked to inline schemas referenced by id', () => {
+  describe('when asked to inline references by id', () => {
     const schemas = [
       {
         $id: '#myDetails',
@@ -247,6 +247,25 @@ describe('Given a utility service for handling dynamic layouts', () => {
       },
     ];
 
+    const actions = [
+        {
+          $id: "#submitMyDetails",
+          title: "Continue",
+          url: "/v3",
+          type: "primary",
+          method: "POST",
+          disabled: false
+        },
+        {
+          $id: "#submitMyAddress",
+          title: "Continue",
+          url: "/v3",
+          type: "primary",
+          method: "POST",
+          disabled: false
+        }
+    ]
+
     it('should inline schema of top level form components', () => {
       const simpleLayout = [
         {
@@ -260,6 +279,14 @@ describe('Given a utility service for handling dynamic layouts', () => {
           schema: {
             $ref: '#myAddress',
           },
+        },
+        {
+          type: 'action',
+          $ref: '#submitMyDetails',
+        },
+        {
+          type: 'action',
+          $ref: '#submitMyAddress',
         },
       ];
 
@@ -284,9 +311,31 @@ describe('Given a utility service for handling dynamic layouts', () => {
             },
           },
         },
+        {
+          type: "button",
+          context: "primary",
+          action: {
+            $id: "#submitMyDetails",
+            label: "Continue",
+            url: "/v3",
+            method: "POST",
+            disabled: false
+          }
+        },
+        {
+          type: "button",
+          context: "primary",
+          action: {
+            $id: "#submitMyAddress",
+            label: "Continue",
+            url: "/v3",
+            method: "POST",
+            disabled: false
+          }
+        }
       ];
 
-      expect(inlineFormSchemas(simpleLayout, schemas)).toEqual(expected);
+      expect(inlineReferences(simpleLayout, schemas, actions)).toEqual(expected);
     });
 
     it('should inline schemas inside boxes', () => {
@@ -300,6 +349,10 @@ describe('Given a utility service for handling dynamic layouts', () => {
                 $ref: '#myAddress',
               },
             },
+            {
+              type: 'action',
+              $ref: '#submitMyAddress',
+            }
           ],
         },
       ];
@@ -318,11 +371,22 @@ describe('Given a utility service for handling dynamic layouts', () => {
                 },
               },
             },
+            {
+              type: "button",
+              context: "primary",
+              action: {
+                $id: "#submitMyAddress",
+                label: "Continue",
+                url: "/v3",
+                method: "POST",
+                disabled: false
+              }
+            },
           ],
         },
       ];
 
-      expect(inlineFormSchemas(boxLayout, schemas)).toEqual(expected);
+      expect(inlineReferences(boxLayout, schemas, actions)).toEqual(expected);
     });
 
     it('should inline schemas inside columns', () => {
@@ -336,6 +400,10 @@ describe('Given a utility service for handling dynamic layouts', () => {
                 $ref: '#myDetails',
               },
             },
+            {
+              type: 'action',
+              $ref: '#submitMyDetails',
+            }
           ],
           right: [
             {
@@ -344,6 +412,10 @@ describe('Given a utility service for handling dynamic layouts', () => {
                 $ref: '#myAddress',
               },
             },
+            {
+              type: 'action',
+              $ref: '#submitMyAddress',
+            }
           ],
         },
       ];
@@ -362,6 +434,17 @@ describe('Given a utility service for handling dynamic layouts', () => {
                 },
               },
             },
+            {
+              type: "button",
+              context: "primary",
+              action: {
+                $id: "#submitMyDetails",
+                label: "Continue",
+                url: "/v3",
+                method: "POST",
+                disabled: false
+              }
+            },
           ],
           right: [
             {
@@ -374,11 +457,22 @@ describe('Given a utility service for handling dynamic layouts', () => {
                 },
               },
             },
+            {
+              type: "button",
+              context: "primary",
+              action: {
+                $id: "#submitMyAddress",
+                label: "Continue",
+                url: "/v3",
+                method: "POST",
+                disabled: false
+              }
+            },
           ],
         },
       ];
 
-      expect(inlineFormSchemas(columnLayout, schemas)).toEqual(expected);
+      expect(inlineReferences(columnLayout, schemas, actions)).toEqual(expected);
     });
   });
 
@@ -395,15 +489,15 @@ describe('Given a utility service for handling dynamic layouts', () => {
     ];
 
     it('should return the original layout', () => {
-      expect(inlineFormSchemas(layout, undefined)).toEqual(layout);
-      expect(inlineFormSchemas(layout, [])).toEqual(layout);
+      expect(inlineReferences(layout, undefined)).toEqual(layout);
+      expect(inlineReferences(layout, [])).toEqual(layout);
     });
   });
 
   describe('when asked to inline schemas and there is no layout', () => {
     it('should return an empty layout', () => {
-      expect(inlineFormSchemas(undefined, undefined)).toEqual([]);
-      expect(inlineFormSchemas(undefined, [])).toEqual([]);
+      expect(inlineReferences(undefined, undefined)).toEqual([]);
+      expect(inlineReferences(undefined, [])).toEqual([]);
     });
   });
 });