+++ packages/dynamic-flows/src/flow/layoutService/index.js
@@ -1,1 +1,1 @@
-export { convertStepToLayout, inlineFormSchemas } from './LayoutService';
+export { convertStepToLayout, inlineReferences } from './LayoutService';