+++ packages/docs/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [11.3.1](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@11.3.0...@transferwise/neptune-docs@11.3.1) (2021-04-08)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+# [11.3.0](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@11.2.5...@transferwise/neptune-docs@11.3.0) (2021-04-08)
+
+
+### Features
+
+* add logo component ([#885](https://github.com/transferwise/neptune-web/issues/885)) ([0110516](https://github.com/transferwise/neptune-web/commit/0110516d85a551956095b44675dae15a581b594e))
+
+
+
+
+
 ## [11.2.5](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@11.2.4...@transferwise/neptune-docs@11.2.5) (2021-04-07)